+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/member/domain/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.nextdoor.nextdoor.domain.member.repository;
+package com.nextdoor.nextdoor.domain.member.domain.repository;
 
 import com.nextdoor.nextdoor.domain.member.domain.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
