+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/query/MemberQueryAdapter.java
@@ -4,7 +4,7 @@ import com.nextdoor.nextdoor.common.Adapter;
 import com.nextdoor.nextdoor.domain.fintech.port.MemberQueryPort;
 import com.nextdoor.nextdoor.domain.member.domain.model.Member;
 import com.nextdoor.nextdoor.domain.member.domain.model.QMember;
-import com.nextdoor.nextdoor.domain.member.repository.MemberRepository;
+import com.nextdoor.nextdoor.domain.member.domain.repository.MemberRepository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
